Cp-003-Map-v2: Fixed Map.vue component path. (all work)
